backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
sqlalchemy==2.0.26
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
